init_db.py
from sqlalchemy import create_engine
from sqlalchemy.orm import sessionmaker
from database import Base
import models
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Get database URL from environment variable
DATABASE_URL = os.getenv("DATABASE_URL")

def init_db():
    # Create SQLAlchemy engine
    engine = create_engine(DATABASE_URL)
    
    # Create all tables
    Base.metadata.create_all(bind=engine)
    
    print("Database tables created successfully!")

if __name__ == "__main__":
    init_db()
